Refactor FHIR version workflows
