+++ src/generator/r4b.ts
@@ -1,36 +1,13 @@
-import { resolve } from "node:path";
-import { writeNormalizedZodDefinitions } from "./emit/zod.ts";
-import { repoRoot } from "./shared.ts";
-import { buildStructureDefinitionR4BDefinitions } from "./sources/structuredefinition-r4b.ts";
+import {
+	fhirReleases,
+	type GenerateVersionOptions,
+	type GenerationResult,
+} from "./versions.ts";
 
-type GenerationResult = {
-	files: string[];
-};
-
-export type GenerateR4BOptions = {
-	generatedAt?: string;
-	outputDir?: string;
-	prune?: boolean;
-	scopeNames?: Iterable<string>;
-};
+export type GenerateR4BOptions = GenerateVersionOptions;
 
 export function generateR4B(
 	options: GenerateR4BOptions = {},
 ): GenerationResult {
-	const structureDefinitionResult = buildStructureDefinitionR4BDefinitions(
-		options.scopeNames,
-	);
-	const outputDir = options.outputDir ?? resolve(repoRoot, "src", "r4b");
-	const generatedAt = options.generatedAt ?? new Date().toISOString();
-	const files = writeNormalizedZodDefinitions({
-		definitions: structureDefinitionResult.definitions,
-		generatedAt,
-		outputDir,
-		prune: options.prune ?? true,
-		primitivePatterns: structureDefinitionResult.primitivePatterns,
-	});
-
-	return {
-		files,
-	};
+	return fhirReleases.r4b.generate(options);
 }
