+++ src/generator/sources/structuredefinition-stu3.ts
@@ -1,15 +1,8 @@
-import { listSTU3GenerationTargetNames } from "../targets/stu3.ts";
-import {
-	buildStructureDefinitionsFromSpec,
-	type StructureDefinitionBuildResult,
-} from "./structuredefinition-r4.ts";
+import { fhirReleases } from "../versions.ts";
+import type { StructureDefinitionBuildResult } from "./structuredefinition.ts";
 
 export function buildStructureDefinitionSTU3Definitions(
-	scopeNames: Iterable<string> = listSTU3GenerationTargetNames(),
+	scopeNames?: Iterable<string>,
 ): StructureDefinitionBuildResult {
-	return buildStructureDefinitionsFromSpec({
-		releaseLabel: "STU3",
-		scopeNames,
-		version: "stu3",
-	});
+	return fhirReleases.stu3.buildDefinitions(scopeNames);
 }
