+++ src/generator/stu3.ts
@@ -1,36 +1,13 @@
-import { resolve } from "node:path";
-import { writeNormalizedZodDefinitions } from "./emit/zod.ts";
-import { repoRoot } from "./shared.ts";
-import { buildStructureDefinitionSTU3Definitions } from "./sources/structuredefinition-stu3.ts";
+import {
+	fhirReleases,
+	type GenerateVersionOptions,
+	type GenerationResult,
+} from "./versions.ts";
 
-type GenerationResult = {
-	files: string[];
-};
-
-export type GenerateSTU3Options = {
-	generatedAt?: string;
-	outputDir?: string;
-	prune?: boolean;
-	scopeNames?: Iterable<string>;
-};
+export type GenerateSTU3Options = GenerateVersionOptions;
 
 export function generateSTU3(
 	options: GenerateSTU3Options = {},
 ): GenerationResult {
-	const structureDefinitionResult = buildStructureDefinitionSTU3Definitions(
-		options.scopeNames,
-	);
-	const outputDir = options.outputDir ?? resolve(repoRoot, "src", "stu3");
-	const generatedAt = options.generatedAt ?? new Date().toISOString();
-	const files = writeNormalizedZodDefinitions({
-		definitions: structureDefinitionResult.definitions,
-		generatedAt,
-		outputDir,
-		prune: options.prune ?? true,
-		primitivePatterns: structureDefinitionResult.primitivePatterns,
-	});
-
-	return {
-		files,
-	};
+	return fhirReleases.stu3.generate(options);
 }
