+++ src/generator/targets/stu3.ts
@@ -1,142 +1,31 @@
-import { readdirSync, readFileSync } from "node:fs";
-import { join } from "node:path";
-import { resolveRequiredSpecPackageRoot } from "../../spec/spec-cache.ts";
+import {
+	fhirReleases,
+	type TargetCategory,
+	type TargetEntry,
+	type TargetSummary,
+} from "../versions.ts";
 
-type StructureDefinition = {
-	abstract?: boolean;
-	baseDefinition?: string;
-	kind?: string;
-	name: string;
-	resourceType: "StructureDefinition";
-	type?: string;
-	url?: string;
-};
+export type STU3TargetCategory = TargetCategory;
+export type STU3TargetEntry = TargetEntry;
 
-export type STU3TargetCategory =
-	| "abstract-whitelist"
-	| "core-resource"
-	| "other"
-	| "profile-resource";
+const release = fhirReleases.stu3;
 
-export type STU3TargetEntry = {
-	abstract: boolean;
-	baseDefinition: string | null;
-	category: STU3TargetCategory;
-	kind: string | null;
-	name: string;
-	shouldGenerate: boolean;
-	type: string | null;
-	url: string | null;
-};
-
-export const stu3AbstractTargetNames = [
-	"BackboneElement",
-	"DomainResource",
-	"Element",
-	"Resource",
-] as const;
-
-const abstractGenerationWhitelist = new Set<string>(stu3AbstractTargetNames);
+export const stu3AbstractTargetNames = release.abstractTargetNames;
 
 export function loadSTU3TargetEntries(): STU3TargetEntry[] {
-	const packageRoot = resolveRequiredSpecPackageRoot("stu3");
-
-	return readdirSync(packageRoot)
-		.filter(
-			(filename) =>
-				filename.startsWith("StructureDefinition-") &&
-				filename.endsWith(".json"),
-		)
-		.map(
-			(filename) =>
-				JSON.parse(
-					stripJsonBom(readFileSync(join(packageRoot, filename), "utf8")),
-				) as StructureDefinition,
-		)
-		.filter((definition) => definition.resourceType === "StructureDefinition")
-		.map((definition) => classifyTargetEntry(definition))
-		.sort((left, right) => left.name.localeCompare(right.name));
-}
-
-function stripJsonBom(content: string): string {
-	return content.charCodeAt(0) === 0xfeff ? content.slice(1) : content;
+	return release.loadTargetEntries();
 }
 
 export function listSTU3CoreResourceNames(): string[] {
-	return loadSTU3TargetEntries()
-		.filter((entry) => entry.category === "core-resource")
-		.map((entry) => entry.name);
+	return release.listCoreResourceNames();
 }
 
 export function listSTU3GenerationTargetNames(): string[] {
-	return loadSTU3TargetEntries()
-		.filter((entry) => entry.shouldGenerate)
-		.map((entry) => entry.name);
+	return release.listGenerationTargetNames();
 }
 
-export function summarizeSTU3Targets(entries = loadSTU3TargetEntries()): {
-	abstractGenerationWhitelist: string[];
-	concreteResourceCount: number;
-	coreResourceCount: number;
-	generationTargetCount: number;
-	profileResourceCount: number;
-} {
-	const concreteResourceCount = entries.filter(
-		(entry) =>
-			entry.kind === "resource" &&
-			entry.abstract === false &&
-			(entry.category === "core-resource" ||
-				entry.category === "profile-resource"),
-	).length;
-	const coreResourceCount = entries.filter(
-		(entry) => entry.category === "core-resource",
-	).length;
-	const generationTargetCount = entries.filter(
-		(entry) => entry.shouldGenerate,
-	).length;
-	const profileResourceCount = entries.filter(
-		(entry) => entry.category === "profile-resource",
-	).length;
-
-	return {
-		abstractGenerationWhitelist: [...abstractGenerationWhitelist].sort(),
-		concreteResourceCount,
-		coreResourceCount,
-		generationTargetCount,
-		profileResourceCount,
-	};
-}
-
-function classifyTargetEntry(definition: StructureDefinition): STU3TargetEntry {
-	const isWhitelistedAbstract = abstractGenerationWhitelist.has(
-		definition.name,
-	);
-	const isConcreteResource =
-		definition.kind === "resource" && definition.abstract !== true;
-	const isCoreResource =
-		isConcreteResource &&
-		definition.type === definition.name &&
-		(definition.baseDefinition ===
-			"http://hl7.org/fhir/StructureDefinition/Resource" ||
-			definition.baseDefinition ===
-				"http://hl7.org/fhir/StructureDefinition/DomainResource");
-	const isProfileResource = isConcreteResource && !isCoreResource;
-	const category: STU3TargetEntry["category"] = isWhitelistedAbstract
-		? "abstract-whitelist"
-		: isCoreResource
-			? "core-resource"
-			: isProfileResource
-				? "profile-resource"
-				: "other";
-
-	return {
-		abstract: definition.abstract === true,
-		baseDefinition: definition.baseDefinition ?? null,
-		category,
-		kind: definition.kind ?? null,
-		name: definition.name,
-		shouldGenerate: isCoreResource || isWhitelistedAbstract,
-		type: definition.type ?? null,
-		url: definition.url ?? null,
-	};
+export function summarizeSTU3Targets(
+	entries?: STU3TargetEntry[],
+): TargetSummary {
+	return release.summarizeTargets(entries);
 }
