+++ tests/helpers/require-r4b-spec.ts
@@ -1,27 +1,8 @@
-import {
-	MissingSpecPackageError,
-	resolveRequiredSpecPackageRoot,
-} from "../../src/spec/spec-cache.ts";
+import { getSpecAvailability } from "./spec-availability.ts";
 
 export function getR4BSpecAvailability(): {
 	available: boolean;
 	reason: string | null;
 } {
-	try {
-		resolveRequiredSpecPackageRoot("r4b");
-
-		return {
-			available: true,
-			reason: null,
-		};
-	} catch (error) {
-		if (error instanceof MissingSpecPackageError) {
-			return {
-				available: false,
-				reason: error.message,
-			};
-		}
-
-		throw error;
-	}
+	return getSpecAvailability("r4b");
 }
